feature-FINISH
+ Dynamic decomposed as custom hook `useObserver`
+ Added limit configuration
